backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pdfplumber>=0.10.3
lxml>=5.1.0
google-genai>=1.0.0
python-dotenv>=1.0.1
pypdf>=5.0.0
nltk>=3.9
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
defusedxml>=0.7.1
